+++ libsolidity/codegen/Compiler.cpp
@@ -37,14 +37,16 @@ void Compiler::compileContract(
 	bytes const& _metadata
 )
 {
-	auto static notTransient = [](VariableDeclaration const* _varDeclaration) {
+	auto static notTransientReferenceType = [](VariableDeclaration const* _varDeclaration) {
 		solAssert(_varDeclaration);
-		return _varDeclaration->referenceLocation() != VariableDeclaration::Location::Transient;
+		return
+			_varDeclaration->referenceLocation() != VariableDeclaration::Location::Transient ||
+			!_varDeclaration->hasReferenceOrMappingType();
 	};
 
 	solUnimplementedAssert(
-		ranges::all_of(_contract.stateVariables(), notTransient),
-		"Transient storage variables are not supported."
+		ranges::all_of(_contract.stateVariables(), notTransientReferenceType),
+		"Transient storage reference type variables are not supported."
 	);
 
 	ContractCompiler runtimeCompiler(nullptr, m_runtimeContext, m_optimiserSettings);