+++ test/libsolidity/syntaxTests/variableDeclaration/state_variables_both_transient_storage_identical_name.sol
@@ -1,0 +1,6 @@
+contract C {
+    uint public x;
+    uint public transient x;
+}
+// ----
+// DeclarationError 2333: (36-59): Identifier already declared.